Support to generate models sending _several_ messages to each remote node
Remote nodes are then only marked active after they have been seeing N
messages, rather than 1.

This is just a workaround to figure out how useful this is. For now,
this is not useful enough to properly implement this. The latter would
require to estimate whether multiple messages to the same node would
result in a lower maximum idle time of any node.