reorder: Allow cross-NUMA links for idle nodes again
This is a work in progress, and only reorders one single message.
To reorder more, we need to do this recursively until there is no
more benefit. For that, we have to update the two new lists that
we introduced with updated idle and active times for updated
edges in the output graph