Limit the number of leaf reported to ten
For that, we select the ten latest leaves to receive the message and
report that to the requester.

This fixes a problem where the benchmark scripts where running very
long for some topologies (such as sequential), since the benchmark
had to be evaluated for each leaf node (i.e. up to 63 times on some
machines).